requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
drf-yasg
inflection
packaging
pillow
PyJWT
pytz
PyYAML
sqlparse
tzdata
uritemplate
